2022/30-陈龙腾 指导老师-陈小军/SCAN/requirements1.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2021.10.8
charset-normalizer==3.0.1
conda==4.3.16
contourpy==1.0.7
cycler==0.11.0
easydict==1.9
faiss-gpu==1.6.4
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
numpy==1.21.4
oauthlib==3.2.2
packaging==23.0
pandas==1.3.4
Pillow==8.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.0.0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchnet==0.0.4
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
typing_extensions==4.0.0
urllib3==1.26.14
visdom==0.2.3
websocket-client==1.5.1
Werkzeug==2.2.2
zipp==3.12.1
